basic test stuff

+++ src/test/java/com/prairiegrade/ugly/Serializer.java
@@ -1,6 +1,8 @@
 package com.prairiegrade.ugly;
 
 import java.io.IOException;
+import java.io.OutputStream;
+import java.io.Writer;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -9,14 +11,16 @@ import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.prairiegrade.ugly.entity.Account;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Dumps some rows into the database, which we can use for testing in our unit
- * test.
+ * Dumps all the {@link Account} entries from the database
+ * to the "data" directory.
  */
 public class Serializer {
     private static final Logger logger = LoggerFactory.getLogger(Serializer.class);
@@ -28,21 +32,21 @@ public class Serializer {
         // create a place to store all this test data
         Path dir = Paths.get("").resolve("data");
         Files.createDirectories(dir);
-
+        GsonBuilder builder = new GsonBuilder().excludeFieldsWithoutExposeAnnotation();
+        Gson gson = builder.create();
         try {
-            TestHarness harness = new TestHarness(em);
-            em.createQuery("from Account", Account.class).getResultList().forEach(account -> {
+            for (Account account : em.createQuery("from Account", Account.class).getResultList()){
                 String filename = String.format("%s.%s.json", account.getClass().getSimpleName(), account.getId());
-
                 Path file = dir.resolve(filename);
-                try {
+                if(Files.exists(file)){
                     Files.delete(file);
-                    harness.transform(Account.class, account.getId(), System.out);
-                } catch(IOException e){
-                    
                 }
-            });
-
+                Files.createFile(file);
+                try(Writer writer = Files.newBufferedWriter(file)){
+                    gson.toJson(account, writer);
+                    logger.debug("Wrote {}#{} to {}", account.getClass(), account.getId(), file);
+                }
+            }
             logger.debug("Done inserting, cleaning up...");
         } finally {
             em.close();
